Move helper function to context.py
- The `ctx.read += 1` now happens in the helper function
- Renamed ctx.excluded to ctx.unmatched
