+++ flow/record/adapter/csvfile.py
@@ -9,8 +9,8 @@ from typing import TYPE_CHECKING
 from flow.record import RecordDescriptor
 from flow.record.adapter import AbstractReader, AbstractWriter
 from flow.record.base import Record, normalize_fieldname
-from flow.record.context import get_app_context
-from flow.record.selector import make_selector, match_record_with_context
+from flow.record.context import get_app_context, match_record_with_context
+from flow.record.selector import make_selector
 from flow.record.utils import boolean_argument, is_stdout
 
 if TYPE_CHECKING:
@@ -120,6 +120,5 @@ class CsvfileReader(AbstractReader):
         for row in self.reader:
             rdict = dict(zip(self.fields, row))
             record = self.desc.init_from_dict(rdict)
-            ctx.read += 1
             if match_record_with_context(record, selector, ctx):
                 yield record
