+++ flow/record/adapter/jsonfile.py
@@ -6,9 +6,9 @@ from typing import TYPE_CHECKING, BinaryIO
 from flow import record
 from flow.record import JsonRecordPacker
 from flow.record.adapter import AbstractReader, AbstractWriter
-from flow.record.context import get_app_context
+from flow.record.context import get_app_context, match_record_with_context
 from flow.record.fieldtypes import fieldtype_for_value
-from flow.record.selector import make_selector, match_record_with_context
+from flow.record.selector import make_selector
 from flow.record.utils import boolean_argument, is_stdout
 
 if TYPE_CHECKING:
@@ -81,7 +81,6 @@ class JsonfileReader(AbstractReader):
         for line in self.fp:
             obj = self.packer.unpack(line)
             if isinstance(obj, record.Record):
-                ctx.read += 1
                 if match_record_with_context(obj, selector, ctx):
                     yield obj
             elif isinstance(obj, record.RecordDescriptor):
@@ -94,6 +93,5 @@ class JsonfileReader(AbstractReader):
                 ]
                 desc = record.RecordDescriptor("json/record", fields)
                 obj = desc(**jd)
-                ctx.read += 1
                 if match_record_with_context(obj, selector, ctx):
                     yield obj
