+++ flow/record/adapter/sqlite.py
@@ -9,8 +9,8 @@ from typing import TYPE_CHECKING
 from flow.record import Record, RecordDescriptor
 from flow.record.adapter import AbstractReader, AbstractWriter
 from flow.record.base import RESERVED_FIELDS, normalize_fieldname
-from flow.record.context import get_app_context
-from flow.record.selector import Selector, make_selector, match_record_with_context
+from flow.record.context import get_app_context, match_record_with_context
+from flow.record.selector import Selector, make_selector
 
 if TYPE_CHECKING:
     from collections.abc import Iterator
@@ -201,7 +201,6 @@ class SqliteReader(AbstractReader):
         for table_name in self.table_names():
             self.logger.debug("Reading table: %s", table_name)
             for record in self.read_table(table_name):
-                ctx.read += 1
                 if match_record_with_context(record, selector, ctx):
                     yield record
