+++ tests/record/test_context.py
@@ -10,7 +10,7 @@ def test_record_context() -> None:
     ctx = get_app_context()
     assert ctx.read == 0
     assert ctx.matched == 0
-    assert ctx.excluded == 0
+    assert ctx.unmatched == 0
 
 
 def test_record_context_metrics(tmp_path: Path) -> None:
@@ -23,12 +23,12 @@ def test_record_context_metrics(tmp_path: Path) -> None:
 
     assert ctx.read == 0
     assert ctx.matched == 0
-    assert ctx.excluded == 0
+    assert ctx.unmatched == 0
 
     list(RecordReader(tmp_path / "test.records", selector="r.number % 2 == 0 or r.number < 1337"))
     assert ctx.read == 2000
     assert ctx.matched == 1668
-    assert ctx.excluded == 332
+    assert ctx.unmatched == 332
 
 
 def test_fresh_app_context(tmp_path: Path) -> None:
@@ -40,27 +40,27 @@ def test_fresh_app_context(tmp_path: Path) -> None:
 
     assert ctx.read == 0
     assert ctx.matched == 0
-    assert ctx.excluded == 0
+    assert ctx.unmatched == 0
 
     list(RecordReader(tmp_path / "test.records", selector="r.number % 2 == 0 or r.number < 1337"))
     assert ctx.read == 2000
     assert ctx.matched == 1668
-    assert ctx.excluded == 332
+    assert ctx.unmatched == 332
 
     with fresh_app_context() as new_ctx:
         assert new_ctx.read == 0
         list(RecordReader(tmp_path / "test.records", selector="r.number == 42"))
         assert new_ctx.read == 2000
         assert new_ctx.matched == 1
-        assert new_ctx.excluded == 1999
+        assert new_ctx.unmatched == 1999
 
     # check if the old context still holds
     assert ctx.read == 2000
     assert ctx.matched == 1668
-    assert ctx.excluded == 332
+    assert ctx.unmatched == 332
 
     # check if the old context still holds via get_app_context()
     ctx = get_app_context()
     assert ctx.read == 2000
     assert ctx.matched == 1668
-    assert ctx.excluded == 332
+    assert ctx.unmatched == 332
